requirements.txt
alembic==1.6.0
attrs==21.2.0
click==7.1.2
Flask==1.1.2
greenlet==1.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.13
SQLAlchemy-Utils==0.37.2
toml==0.10.2
Werkzeug==1.0.1
